Move driver code to library
