+++ examples/rbfe_edge_list.py
@@ -1,85 +1,6 @@
 import argparse
-import csv
-import pickle
-import traceback
 
-import numpy as np
-from rdkit import Chem
-
-Chem.SetDefaultPickleProperties(Chem.PropertyPickleOptions.AllProps)
-
-from simtk.openmm import app
-
-from timemachine.constants import KCAL_TO_KJ
-from timemachine.fe import atom_mapping
-from timemachine.fe.rbfe import run_complex, run_solvent
-from timemachine.fe.utils import get_mol_name, read_sdf
-from timemachine.ff import Forcefield
-from timemachine.parallel.client import CUDAPoolClient
-
-
-def get_mol_by_name(mols, name):
-    for m in mols:
-        if get_mol_name(m) == name:
-            return m
-
-    assert 0, "Mol not found"
-
-
-def run_edge_and_save_results(
-    mol_a,
-    mol_b,
-    core,
-    forcefield,
-    protein,
-    n_frames,
-    seed,
-    smarts,
-    exp_ddg,
-    fep_ddg,
-    fep_ddg_err,
-    ccc_ddg,
-    ccc_ddg_err,
-):
-    mol_a_name = get_mol_name(mol_a)
-    mol_b_name = get_mol_name(mol_b)
-
-    try:
-        complex_res, complex_top = run_complex(mol_a, mol_b, core, forcefield, protein, n_frames, seed)
-        solvent_res, solvent_top = run_solvent(mol_a, mol_b, core, forcefield, protein, n_frames, seed)
-
-        meta = (
-            mol_a,
-            mol_b,
-            smarts,
-            core,
-            float(exp_ddg) * KCAL_TO_KJ,
-            float(fep_ddg) * KCAL_TO_KJ,
-            float(fep_ddg_err) * KCAL_TO_KJ,
-            float(ccc_ddg) * KCAL_TO_KJ,
-            float(ccc_ddg_err) * KCAL_TO_KJ,
-        )
-
-        with open(f"success_rbfe_result_{mol_a_name}_{mol_b_name}.pkl", "wb") as fh:
-            pkl_obj = (meta, solvent_res, solvent_top, complex_res, complex_top)
-            pickle.dump(pkl_obj, fh)
-
-        solvent_ddg = np.sum(solvent_res.all_dGs)
-        solvent_ddg_err = np.linalg.norm(solvent_res.all_errs)
-        complex_ddg = np.sum(complex_res.all_dGs)
-        complex_ddg_err = np.linalg.norm(complex_res.all_errs)
-
-        tm_ddg = complex_ddg - solvent_ddg
-        tm_err = np.linalg.norm([complex_ddg_err, solvent_ddg_err])
-
-        print(
-            f"finished: {mol_a_name} -> {mol_b_name} (kJ/mol) | complex {complex_ddg:.2f} +- {complex_ddg_err:.2f} | solvent {solvent_ddg:.2f} +- {solvent_ddg_err:.2f} | tm_pred {tm_ddg:.2f} +- {tm_err:.2f} | exp_ddg {exp_ddg:.2f} | fep_ddg {fep_ddg:.2f} +- {fep_ddg_err:.2f}"
-        )
-    except Exception as err:
-        print(
-            f"failed: {err} {mol_a_name} -> {mol_b_name} (kJ/mol) | exp_ddg {exp_ddg:.2f} | fep_ddg {fep_ddg:.2f} +- {fep_ddg_err:.2f}"
-        )
-        traceback.print_exc()
+from timemachine.fe import rbfe
 
 
 def parse_args():
@@ -100,56 +21,9 @@ def parse_args():
     return parser.parse_args()
 
 
-def run_parallel(n_frames, ligands, results_csv, forcefield, protein, n_gpus, seed):
-    mols = read_sdf(str(ligands))
-
-    cpc = CUDAPoolClient(n_gpus)
-    cpc.verify()
-
-    ff = Forcefield.load_from_file(forcefield)
-    protein = app.PDBFile(protein)
-
-    with open(results_csv) as csvfile:
-        reader = csv.reader(csvfile, delimiter=",")
-        next(reader)
-        rows = [row for row in reader]
-        futures = []
-        for row_idx, row in enumerate(rows):
-            mol_a_name, mol_b_name, exp_ddg, fep_ddg, fep_ddg_err, ccc_ddg, ccc_ddg_err = row
-            mol_a = get_mol_by_name(mols, mol_a_name)
-            mol_b = get_mol_by_name(mols, mol_b_name)
-
-            mcs_threshold = 2.0
-            core, smarts = atom_mapping.get_core_with_alignment(mol_a, mol_b, threshold=mcs_threshold)
-
-            print(f"Submitting job for {mol_a_name} -> {mol_b_name}")
-            futures.append(
-                cpc.submit(
-                    run_edge_and_save_results,
-                    mol_a,
-                    mol_b,
-                    core,
-                    forcefield,
-                    protein,
-                    args.n_frames,
-                    args.seed + row_idx,
-                    smarts,
-                    exp_ddg,
-                    fep_ddg,
-                    fep_ddg_err,
-                    ccc_ddg,
-                    ccc_ddg_err,
-                )
-            )
-
-        # Block until subprocesses finish (possibly redundant; included to be explicit)
-        for future in futures:
-            _ = future.result()
-
-
 if __name__ == "__main__":
     args = parse_args()
-    run_parallel(
+    rbfe.run_parallel(
         args.n_frames,
         args.ligands,
         args.results_csv,
