+++ timemachine/fe/rbfe.py
@@ -1,22 +1,30 @@
+import csv
 import functools
 import io
 import pickle
+import traceback
 import warnings
+from pathlib import Path
+from typing import Union
 
 import matplotlib.pyplot as plt
 import numpy as np
 import pymbar
+from rdkit import Chem
+from simtk.openmm import app
 
-from timemachine.constants import BOLTZ, DEFAULT_TEMP
-from timemachine.fe import model_utils
+from timemachine.constants import BOLTZ, DEFAULT_TEMP, KCAL_TO_KJ
+from timemachine.fe import atom_mapping, model_utils
 from timemachine.fe.bar import bar_with_bootstrapped_uncertainty
 from timemachine.fe.free_energy import HostConfig, InitialState, SimulationProtocol, SimulationResult
 from timemachine.fe.single_topology import SingleTopology
 from timemachine.fe.system import convert_omm_system
-from timemachine.fe.utils import get_mol_name, get_romol_conf
+from timemachine.fe.utils import get_mol_name, get_romol_conf, read_sdf
+from timemachine.ff import Forcefield
 from timemachine.lib import LangevinIntegrator, MonteCarloBarostat, custom_ops
 from timemachine.md import builders, minimizer
 from timemachine.md.barostat.utils import get_bond_list, get_group_indices
+from timemachine.parallel.client import CUDAPoolClient
 from timemachine.potentials import jax_utils
 
 
@@ -679,3 +687,119 @@ def run_complex(
         steps_per_frame=steps_per_frame,
     )
     return complex_res, complex_top
+
+
+def run_edge_and_save_results(
+    mol_a,
+    mol_b,
+    core,
+    forcefield,
+    protein,
+    n_frames,
+    seed,
+    smarts,
+    exp_ddg,
+    fep_ddg,
+    fep_ddg_err,
+    ccc_ddg,
+    ccc_ddg_err,
+):
+    mol_a_name = get_mol_name(mol_a)
+    mol_b_name = get_mol_name(mol_b)
+
+    try:
+        complex_res, complex_top = run_complex(mol_a, mol_b, core, forcefield, protein, n_frames, seed)
+        solvent_res, solvent_top = run_solvent(mol_a, mol_b, core, forcefield, protein, n_frames, seed)
+
+        meta = (
+            mol_a,
+            mol_b,
+            smarts,
+            core,
+            float(exp_ddg) * KCAL_TO_KJ,
+            float(fep_ddg) * KCAL_TO_KJ,
+            float(fep_ddg_err) * KCAL_TO_KJ,
+            float(ccc_ddg) * KCAL_TO_KJ,
+            float(ccc_ddg_err) * KCAL_TO_KJ,
+        )
+
+        with open(f"success_rbfe_result_{mol_a_name}_{mol_b_name}.pkl", "wb") as fh:
+            pkl_obj = (meta, solvent_res, solvent_top, complex_res, complex_top)
+            pickle.dump(pkl_obj, fh)
+
+        solvent_ddg = np.sum(solvent_res.all_dGs)
+        solvent_ddg_err = np.linalg.norm(solvent_res.all_errs)
+        complex_ddg = np.sum(complex_res.all_dGs)
+        complex_ddg_err = np.linalg.norm(complex_res.all_errs)
+
+        tm_ddg = complex_ddg - solvent_ddg
+        tm_err = np.linalg.norm([complex_ddg_err, solvent_ddg_err])
+
+        print(
+            f"finished: {mol_a_name} -> {mol_b_name} (kJ/mol) | complex {complex_ddg:.2f} +- {complex_ddg_err:.2f} | solvent {solvent_ddg:.2f} +- {solvent_ddg_err:.2f} | tm_pred {tm_ddg:.2f} +- {tm_err:.2f} | exp_ddg {exp_ddg:.2f} | fep_ddg {fep_ddg:.2f} +- {fep_ddg_err:.2f}"
+        )
+    except Exception as err:
+        print(
+            f"failed: {err} {mol_a_name} -> {mol_b_name} (kJ/mol) | exp_ddg {exp_ddg:.2f} | fep_ddg {fep_ddg:.2f} +- {fep_ddg_err:.2f}"
+        )
+        traceback.print_exc()
+
+
+def run_parallel(
+    n_frames: int,
+    ligands: Union[Path, str],
+    results_csv: Union[Path, str],
+    forcefield: Union[Path, str],
+    protein_pdb: Union[Path, str],
+    n_gpus: int,
+    seed: int,
+):
+    mols = {get_mol_name(mol): mol for mol in read_sdf(str(ligands))}
+
+    cpc = CUDAPoolClient(n_gpus)
+    cpc.verify()
+
+    ff = Forcefield.load_from_file(forcefield)
+    protein = app.PDBFile(str(protein_pdb))
+
+    # Ensure that all mol props (e.g. _Name) are included in pickles
+    # Without this get_mol_name(mol) will fail on roundtripped mol
+    Chem.SetDefaultPickleProperties(Chem.PropertyPickleOptions.AllProps)
+
+    with open(results_csv) as csvfile:
+        reader = csv.reader(csvfile, delimiter=",")
+        next(reader)
+        rows = [row for row in reader]
+        futures = []
+
+        for row_idx, row in enumerate(rows):
+            mol_a_name, mol_b_name, exp_ddg, fep_ddg, fep_ddg_err, ccc_ddg, ccc_ddg_err = row
+            mol_a = mols[mol_a_name]
+            mol_b = mols[mol_b_name]
+
+            mcs_threshold = 2.0
+            core, smarts = atom_mapping.get_core_with_alignment(mol_a, mol_b, threshold=mcs_threshold)
+
+            print(f"Submitting job for {mol_a_name} -> {mol_b_name}")
+            futures.append(
+                cpc.submit(
+                    run_edge_and_save_results,
+                    mol_a,
+                    mol_b,
+                    core,
+                    ff,
+                    protein,
+                    n_frames,
+                    seed + row_idx,
+                    smarts,
+                    exp_ddg,
+                    fep_ddg,
+                    fep_ddg_err,
+                    ccc_ddg,
+                    ccc_ddg_err,
+                )
+            )
+
+        # Block until subprocesses finish (possibly redundant; included to be explicit)
+        for future in futures:
+            _ = future.result()
